feat: add conditional `rsc` mode support (#21)
* feat: add conditional `rsc` mode support

* chore: add changeset

* refactor(compiler): restructure plugins dir
