+++ .changeset/seven-poems-retire.md
@@ -1,0 +1,5 @@
+---
+"@replexica/compiler": minor
+---
+
+The compiler now supports the rsc flag, which is set to true by default. When set to false, every piece of code is processed as if it had 'use client'.
